[Arch] Extract craft feature into src/features/craft/
Move all craft/playground code from src/components/ into a self-contained
src/features/craft/ directory. Extract CraftListPage and CraftDetailPage
into feature components, making route files thin wrappers that only handle
metadata and static params.

Closes ROB-15
