+++ src/app/craft/page.tsx
@@ -1,11 +1,7 @@
 // src/app/craft/page.tsx
-import Container from "@/shared/layout/container";
-import { Card } from "@/shared/components/card";
+import { CraftListPage } from "@/features/craft";
 import { allCrafts } from "contentlayer/generated";
-import { compareDesc, format } from "date-fns";
-import { ArrowUpRight } from "lucide-react";
-import Image from "next/image";
-import Link from "next/link";
+import { compareDesc } from "date-fns";
 
 export const metadata = {
   title: "Craft | Robert Crocker",
@@ -18,125 +14,5 @@ export default function CraftPage() {
     .filter((craft) => craft.published)
     .sort((a, b) => compareDesc(new Date(a.date), new Date(b.date)));
 
-  return (
-    <Container className="pt-8 pb-20">
-      {/* Page Header */}
-      <header className="mb-8">
-        <Link
-          href="/#craft"
-          className="inline-flex gap-2 items-center mb-6 text-sm font-medium transition-colors text-muted-foreground hover:text-primary"
-        >
-          ← Back to home
-        </Link>
-        <h1 className="mb-4 text-4xl font-bold tracking-tighter sm:text-5xl">
-          Craft
-        </h1>
-        <p className="max-w-2xl text-lg text-muted-foreground">
-          Interactive experiments exploring the invisible details that make
-          interfaces feel right. Each piece includes a live demo and a deep-dive
-          into the how and why.
-        </p>
-      </header>
-
-      {/* Craft Grid */}
-      <div className="grid gap-6 sm:grid-cols-2 lg:grid-cols-3">
-        {crafts.map((craft) => (
-          <Link
-            key={craft.slug}
-            href={craft.url}
-            className="group focus-visible:outline-none"
-          >
-            <Card className="overflow-hidden ring-2 ring-transparent ring-offset-1 transition-all duration-200 focus-within:ring-primary hover:ring-primary">
-              {/* Preview Media */}
-              <div className="overflow-hidden relative w-full aspect-video bg-muted">
-                {craft.video ? (
-                  <>
-                    {craft.poster && (
-                      <img
-                        src={craft.poster}
-                        alt=""
-                        aria-hidden="true"
-                        className="object-cover absolute inset-0 w-full h-full blur-xl scale-110"
-                      />
-                    )}
-                    <video
-                      autoPlay
-                      loop
-                      muted
-                      playsInline
-                      className="object-cover absolute inset-0 w-full h-full"
-                    >
-                      <source
-                        src={craft.video.replace(".mp4", ".webm")}
-                        type="video/webm"
-                      />
-                      <source src={craft.video} type="video/mp4" />
-                    </video>
-                  </>
-                ) : craft.image ? (
-                  <Image
-                    src={craft.image}
-                    alt={craft.title}
-                    fill
-                    className="object-cover transition-transform duration-300 group-hover:scale-105"
-                    sizes="(max-width: 640px) 100vw, (max-width: 1024px) 50vw, 33vw"
-                  />
-                ) : (
-                  <div className="flex h-full w-full items-center justify-center bg-[radial-gradient(hsl(var(--muted-foreground)/0.1)_1px,transparent_1px)] [background-size:20px_20px]">
-                    <span className="text-4xl font-bold text-muted-foreground/20">
-                      {craft.title.charAt(0)}
-                    </span>
-                  </div>
-                )}
-
-                {/* Hover indicator */}
-                <div className="absolute right-3 top-3 rounded-full bg-primary p-1.5 text-primary-foreground opacity-0 shadow backdrop-blur-sm transition-opacity group-hover:opacity-100">
-                  <ArrowUpRight className="w-4 h-4" />
-                </div>
-              </div>
-
-              {/* Content */}
-              <div className="p-4">
-                <div className="flex justify-between items-center mb-2">
-                  <h2 className="text-lg font-semibold transition-colors group-hover:text-primary">
-                    {craft.title}
-                  </h2>
-                  <time
-                    dateTime={craft.date}
-                    className="text-sm text-muted-foreground"
-                  >
-                    {format(new Date(craft.date), "MMM yyyy")}
-                  </time>
-                </div>
-
-                <p className="text-sm line-clamp-2 text-muted-foreground">
-                  {craft.description}
-                </p>
-
-                {/* Tags */}
-                {craft.tags && craft.tags.length > 0 && (
-                  <div className="mt-3 flex flex-wrap gap-1.5">
-                    {craft.tags.slice(0, 3).map((tag) => (
-                      <span
-                        key={tag}
-                        className="rounded bg-secondary px-2 py-0.5 text-xs text-secondary-foreground"
-                      >
-                        {tag}
-                      </span>
-                    ))}
-                  </div>
-                )}
-              </div>
-            </Card>
-          </Link>
-        ))}
-      </div>
-
-      {crafts.length === 0 && (
-        <div className="py-12 text-center text-muted-foreground">
-          <p>Craft pieces coming soon. Check back shortly.</p>
-        </div>
-      )}
-    </Container>
-  );
+  return <CraftListPage crafts={crafts} />;
 }
