+++ src/features/craft/playgrounds/interactive-dot-grid-playground.tsx
@@ -9,7 +9,7 @@ import {
   PlaygroundCode,
   PlaygroundControls,
   RangeControl,
-} from "@/components/playground";
+} from "../playground";
 
 // --- Types ---
 interface Point {
